pyproject.toml
[project]
name = "livery"
version = "0.6.2"
description = "Local AI company orchestration — markdown + git, no database"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "python-frontmatter>=1.1",
]

[project.scripts]
livery = "livery.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["livery"]

[dependency-groups]
dev = [
    "pytest>=8.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
